+++ changelog.md
@@ -18,3 +18,5 @@
 
 # *Md. Masud Rana*
 - Modify get all assessment API
+- Add assessment stats API
+- Modify single assessment API (*A assessment with all question and option or nested question or option*)